Adding text files
Add a set of 10 text files, each of which includes two questions (with
answers) for each of the five topics there are in the game. Each text
file represents a base for every level of difficulty.

Currently the way of assigning unique numbers to the questions is:
A four-digit or a five-digit format, where the first digit (for
level 10, it's the first two digits) represents
the level of difficulty (1-10), the digit of the hundreds represents
the exact topic (1-5: 1 is History, 2 is Geography, 3 is Science,
4 is Literature and 5 is Fun facts), and the ending digits are the
serial number of the question for that topic in that level of
difficulty.
E.g.: The second question for the Science topic with difficulty level
8 is: 8302.
E.g.(*): For difficulty level 10, we have 10101, 10402, etc.
This is not the final form of the text files, so everything, including
the numeration, might be changed.

Add the possibility to terminate the program if the User presses "X"
in the primary menu.