Add the functionality to read text; First Stage
Adding the logic I'm going to use to read from text files.
Adding the first set of questions (Level 1).
The user can now choose between the six categories of the game and
play the first level.

Making some major changes to the text files:
- Primarily: changing their address, moving them to the directory
of the .cpp and .h files, because I had forgotten to do it earlier;
- Changing the text syntax a little bit:
* Now the questions start just
with their serial number and a space bar in front of the first letter
of each one;
* There is a line of space between the question and the answers;
* For now, every possible answer ends with space bars, the last one
ends with a tabulation ('\t'). The purpose of the tabulations is to
separate each question.;
* Moving C) and D) answers to the first two rows so it looks better;

For now, these changes only apply to the first text file (Level1), the
others I am going to fix later.

Also adding the award screen which informs the user about the current
award and lets them go further in the game, and the defeat screen,
which informs the user about their loss and lets them go back to the
main menu. This kind of screens are going to be made for each level.