The "Phone a friend" and "50/50" lifelines work
Make the functions for "Phone a friend", fix some problems with the
lifelines and include them in the game.
Fix a major problem with the reward screen when a user starts a new
game through the use of a reference of a static variable as a
function parameter.
Lifelines functions are still too long.
Currently adding some of the logic for the "Ask the audience" lifeline.