Game mode is working
Add "Ask the audience" lifeline.
There could be an error with the lifeline I will try to trace later.
All the lifelines are working.
All of the prepared questions (100) are included (with some fixes).
Note: The "Ask the audience" and "Phone a friend" lifelines use
vectors of answers to determine the suggested answer with the necessary
success rate.
E.g.: If we are on 7-10 level of the game (the hard ones), the success
rate of these lifelines must be 20%. So, we take an array with the
possible answers where the correct/wrong answers ratio is
20/100. How do we achieve that? We fill the array with 30 elements,
6 of which are correct (6 is 20% of 30), the other 24 are wrong, 8 of
every type. That means, if A) is correct, we have 6 A's, 8 B's, 8 C's
and 8 D's.
Then we use the rand() function to pick a random element from the array
(string in this case), and this is how the success rate is accomplished.