Add the "Create question" option
Almost all of the "Create" mode functions have been written. The
logic for creating a unique serial number is on and working.
The user should enter a question, then enter the answers, then
choose a right answer, pick a level of difficulty (1-10), and finally
pick a category (History, Geography, Science, Literature, Fun facts).
The unique serial number for the new question is being created like so:
Initially, it's equal to the level of difficulty number, multiplied by
1000, then we sum it up with 100 times the category number (1-5).

E.g.: If the level is 4 and the category is Fun facts, we get
4 * 1000 + 5 * 100, which is equal to 4500. Then we add 1 to it, because
our serial number system does not support numbers that end with 00.
So, we get 4501. Then we iterate through the vector that includes all
of the base questions' serial numbers and check if 4501 is equal to any
of them. In this case, there already is a serial number of 4501, so
we increment our new number until we have no coincidence. For example,
our new number can be 4503.