The "Create a question" option is now working
The logic for the "Create a question" mode is completed, if the user
picks this mode of the game, they can add a question, then,
a unique number is created for this question, which we put in a new
text file, whilst the question itself is being put in one of the 10
text files, responsible for the different levels of difficulty.
Then, when the user accesses the main menu of the game, a vector is
filled with all the current questions and then the program checks if
there are new questions, while reading the AddedQuestionsNumbers.txt
file. That's how we make sure all the questions: the base questions and
our new ones, are included in the game.