Make a question search functionality
There is no need for an array that holds the ready prepared questions
anymore. Now, when we start the game, we check in each one of the 10
level files for questions and then extract their serial numbers and
fill a vector with them. That functionality is triggered every time the
user goes to the main menu. That way, if the user creates a question
through the "Create a question" mode, the question is being saved
and can be presented in the game right away.