Add the "Edit mode" of the game
This mode lets the user find a question by its unique serial number
and edit it. Program-wise, in our case this means that the chosen
question is being read on the console screen, at the same time a copy
of the text file where the question is extracted from is being
created, it includes the very same text as in the original file, but
it excludes the edited question. Then, after the user has successfully
edited the question, it is added in the new file, thus, changing nothing
but the chosen question.
Some edge cases still need fixing, but the game is ready to play with
all of its three modes working.