Fix serial number generator
There has been an issue with generating unique numbers.
Every time one is created, the responsible function would check
in the allQuestions array, where all numbers are included, if there is
already such number. If so, the number would increment. However, this
didn't work if a smaller number is positioned after a bigger number
as an element in the array. Now, the array is being sorted every time
the main menu is presented, to make sure all changes are put
into control.